Move classes for corners into single file
