+++ tests/translate/__init__.py
@@ -12,7 +12,11 @@ from .translate_c_sw import (
     TranslateVorticityTransport_Cgrid,
 )
 from .translate_compute_total_energy import TranslateComputeTotalEnergy
-from .translate_copycorners import TranslateCopyCorners
+from .translate_corners import (
+    TranslateCopyCorners,
+    TranslateFillCorners2D,
+    TranslateFillCornersCells,
+)
 from .translate_cubedtolatlon import TranslateCubedToLatLon
 from .translate_d2a2c_vect import TranslateD2A2C_Vect
 from .translate_d_sw import TranslateD_SW, TranslateUbKE, TranslateVbKE
